Removes old publications
