+++ app/grandchallenge/publications/utils/__init__.py
@@ -1,26 +1,10 @@
-import re
-
 import requests
-from django.db.models import TextChoices
 
-from grandchallenge.publications.models import ARXIV_REGEX, DOI_REGEX
+from grandchallenge.publications.models import (
+    PublicationType,
+    get_publication_type,
+)
 from grandchallenge.publications.utils.manubot import get_arxiv_csl
-
-
-class PublicationType(TextChoices):
-    DOI = "D"
-    ARXIV = "A"
-
-
-def get_publication_type(*, identifier: str) -> PublicationType:
-    if re.match(DOI_REGEX, identifier):
-        return PublicationType.DOI
-    elif re.match(ARXIV_REGEX, identifier):
-        return PublicationType.ARXIV
-    else:
-        raise ValueError(
-            f"Could not determine publication type from {identifier}"
-        )
 
 
 def get_identifier_csl(*, doi_or_arxiv):
